Reorganize and document required prop utils
