+++ tools/analyzer_plugin/lib/src/util/prop_declarations/get_all_props.dart
